Initial version
A basic way to get periodic queries on a database via an Openshift
cronjob.

The Dockerfile is based on ubi9, which gives us a postgres 13 client.
Unfortunately I haven't been able to find a Fedora image Konflux likes.
This is not ideal, but it works for the moment with the current intended
user of these queries.

The plan is to deploy this in a namespace different than the
application owned by us. With that we will be able to make sure that no
crazy queries are run.

APPSRE-11356

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>